// src/movimentacoes/dto/movimentacoes.dto.ts
export interface MovimentacoesDto{
    id: number;
    entrada1: string;
    entrada2: string;
    entrada3: string;
    entrada4: string;
    entrada5: string;
    entrada6: string;
    entrada7: string;
    entrada8: string;
    entrada9: string;
    entrada10: string;
    entrada11: string;
    entrada12: string;
    entrada13: string;
    entrada14: string;
    entrada15: string;
    entrada16: string;
    entrada17: string;
    entrada18: string;
    entrada19: string;
    entrada20: string;
    entrada21: string;
    entrada22: string;
    entrada23: string;
    entrada24: string;
    entrada25: string;
    entrada26: string;
    entrada27: string;
    entrada28: string;
    entrada29: string;
    entrada30: string;
    entrada31: string;
    entrada32: string;
}